Title: 평균점수 구하기, Time: 164ms, Memory: 11MB, Status: Passed - CodeTree

+++ 241009/평균점수 구하기/print-average.java
@@ -3,13 +3,11 @@ import java.util.*;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int sum=0;
-
+        double sum = 0;
         for(int i=0; i<8; i++){
-            int n=sc.nextInt();
+            double n = sc.nextDouble();
             sum+=n;
         }
-
-        System.out.printf("%.1f",(double)sum/8);
+        System.out.printf("%.1f",sum/8);
     }
 }
